# RFC: Commitlog refactor: message loop v2