--- a/db/commitlog/commitlog.cc
+++ b/db/commitlog/commitlog.cc
@@ -923,13 +923,13 @@ public:
     }
 
     // See class comment for info
-    future<sseg_ptr> sync() {
+    future<> sync() {
         // Note: this is not a marker for when sync was finished.
         // It is when it was initiated
         reset_sync_time();
         return cycle(true);
     }
-    future<sseg_ptr> terminate() {
+    future<> terminate() {
         assert(_closed);
         bool terminate = false;
 
@@ -960,9 +960,7 @@ public:
 
         // just in case, wait for flushing to finish up to what is now final end
         // pos.
-        co_return co_await _flush_pos.wait(_file_pos).then([me = shared_from_this()] {
-            return me;
-        });
+        co_await _flush_pos.wait(_file_pos);
     }
     void issue_close_request() {
         if (!std::exchange(_closed, true)) {
@@ -1219,18 +1217,16 @@ public:
      * Send any buffer contents to disk and get a new tmp buffer
      */
     // See class comment for info
-    future<sseg_ptr> cycle(bool flush_after = false, bool termination = false) {
-        auto me = shared_from_this();
+    future<> cycle(bool flush_after = false, bool termination = false) {
         auto top = issue_cycle(flush_after, termination);
         if (flush_after) {
             co_await _flush_pos.when(top);
         } else {
             co_await _written_pos.when(top);
         }
-        co_return me;
     }
 
-    future<sseg_ptr> batch_cycle(timeout_clock::time_point timeout) {
+    future<> batch_cycle(timeout_clock::time_point timeout) {
         /**
          * For batch mode we force a write "immediately".
          * However, we first wait for all previous writes/flushes
@@ -1239,7 +1235,6 @@ public:
          * This has the benefit of allowing several allocations to
          * queue up in a single buffer.
          */
-        auto me = shared_from_this();
         auto fp = _file_pos;
         try {
             // wait for all writes (and maybe flush) to reach our _start_
@@ -1252,7 +1247,7 @@ public:
                 // (Note: wait_for_pending(pos) waits for operation _at_ pos (and before),
                 assert(_segment_manager->cfg.mode != sync_mode::BATCH || _flush_pos > fp);
                 if (_flush_pos > fp) {
-                    co_return me;
+                    co_return;
                     // previous op we were waiting for was not sync one, so it did not flush
                     // force flush here
                 }
@@ -1266,10 +1261,9 @@ public:
             // we should close the segment.
             // TODO: should we also trunctate away any partial write
             // we did?
-            me->_closed = true; // just mark segment as closed, no writes will be done.
+            _closed = true; // just mark segment as closed, no writes will be done.
             throw;
         };
-        co_return me;
     }
 
     void background_cycle() {
@@ -1495,13 +1489,13 @@ future<R> db::commitlog::segment_manager::allocate_when_possible(T writer, db::t
             case write_result::ok:
                 co_return writer.result();
             case write_result::must_sync:
-                s = co_await with_timeout(timeout, s->sync());
+                co_await with_timeout(timeout, s->sync());
                 continue;
             case write_result::no_space:
                 s = co_await s->finish_and_get_new(timeout);
                 continue;
             case write_result::ok_need_batch_sync:
-                s = co_await s->batch_cycle(timeout);
+                co_await s->batch_cycle(timeout);
                 co_return writer.result();
         }
     }