A new take on trying to move operations in commit log into a "message loop" construct, i.e. "requests/tasks", but now split into two distinct execution fibers: "controller" and "writing". Also, now all operations (except shutdown) are executed fully serialized. I.e. we finish message A before getting to message B. 

This means theoretically that we could utilize slightly less bandwidth for write than existing code, but semi-relevant experiments/observations show that we only really have parallel write executions on segment spin-up and close/termination. 
Thus this part of IO (which is less latency sensitive) is handled by the "main" loop, not the dedicated IO one, which only deals with low-latency write/flush stuff. 

Note this is RFC. Needs rebasing, as it is based on previous refactor branch, and can probably do with a little cleanup. 
The idea is to enable further simplification due to the now more serialized nature of the ops. 

See #9222 for the first version/discussion. 
Requires seastar patch "Add basic_condition_variable templated on clock type"

Significant differences: 
* "message queue" is now just a seastar::circular_buffer + condition variable.
* two "queues": one for latency-sensistive writing, one for periodic flushing, segment managemen, alloc, close, terminate etc.
* All ops fully serialized, except shutdown which is partially parallel, and replenish, which is a "background" fiber still.