@avikivity 2022-03-08
Missing a performance evaluation (perf_simple_query in write mode, batch commitlog, tmpfs).