@elcallio 2022-03-09
Working on evaluating performance from various standpoints. There is still work to do to ensure fast enough. This is RFC just for discussing the general approach (full serialization).