@elcallio 2022-03-29
So, perhaps less than stellar results for benchmarking; Running the perf_commitlog bench with small data input and suitably small variance in segment delays etc, I get consistently lower score for this branch vs. master. 

The apparent reason is that when filling up the segment write buffer (128k), and issuing `background_cycle` (write the data to disk in background), even though one would think that the lions share would be in the uninterruptible (non-continuation) part where we write preamble to the buffer, package it etc, actually sending the work order as a packet via the `message_queue` (i.e. circular buffer + condition_variable) add enough instructions in the client write path (waking up sleeping coroutine) that we get a lower score and higher IPO. If I quickly haxxor it to instead do the IO in a non-achored background future like master, the score gets back on track again. This is a little scary, as doing something that should essentially be "schedule something into the task queue" adds that much (~200IPO), vs. "suspend to background" (which is what doing unachored continuation means).