@avikivity 2022-03-30
That's super surprising. 128k flushes are amortized over many (>1000) commits. We should be able to do what we like there.

Are we suppressing unnecessary wakeups?