@avikivity 2022-03-30
Maybe it's latency not throughput. Launching an I/O with (void)do_something() will queue the I/O when the next quota completes, while waking a fiber can delay it if there are many tasks queued already.

However, that means that in the first case tasks keep accumulating. Does the test wait for all work to complete?