@elcallio 2022-03-30
No, just raw, immediate throughput. So pushing things into the future should be good for numbers. :-) Other than building up memory buffer debt...
IO throughput seems to be on par. But I get consistently ~15% higher instr/op when using queued handoff vs. background (non-achored) write. I can't make sense of it though, doing a circular buffer emplace + a schedule should not be that expensive, esp. since we at least _don't_ do the initial IO write parts...