@avikivity 2022-04-06
> Ok, some sunlight: After updating all competing branches to be (roughly) equally in sync with master, I now get pretty much en par performance on `perf_commitlog` (above), with repeats. Still higher task+alloc+instruction count, which I don't 100% like or understand, but it stands to reason it is due to how IO gets scheduled.

How much higher? numbers.

> There might be more to it though, as there are some discrepancies in how many continuations+wakeups we get in top-level bench routine, suggesting we more often cause delay in `allocate_when_possible`, though measurement there cannot really see anything...
> 
> On a more positive note; I get consistently slightly better throughput on perf_simple_query (write) with the new code, and even more visibly so in `perf_commitlog` in batch mode.
> 
> I've done some tweaks to simplify code, waitable_counter now uses cond-var (without spurious wakeups thanks to `when`), and write path in CL could be further slimline (last patch) once we've serialized all writes.
> 
> So, obviously, since we serialize writes, and the segments that need them, we basically limit the IO throughput of CL to one outstanding write/flush per shard. This is by design here, since in "normal" uses we don't seem to stray too much from that anyway, but it means we are rate-limiting CL further than the IO group does (for better or worse).

We are not. If latency grows, so does the batch. Assuming enough concurrency. How much do you have?

> Now, we will still probably have X writes in total done at all times, since per-shard and all. But it does mean we might not (at least in future?) 100% saturate IO bandwidth. One approach to maybe increase IO load would be to couple it with the "segregation" suggestion, i.e. different CL queues for different CF:s. One could then have separate write workers per queue as well, to perhaps utilize even more bandwidth than today.

Don't need that, we could allow > 1 outstanding writes if we wanted. I think that with sufficient client concurrency it is not needed.