@elcallio 2022-04-26
Scylla-bench (write, 120s) using CL hard limit produces pretty even results between PR and master, with a slight edge for the PR version (though the inherent flakiness of ccm-benchmarking means this should always be treated with a grain of salt).

```
PR:
------------------------------
  
Results
Time (avg):      1m59.996542917s
Total ops:       1992818
Total rows:      1992818
Operations/s:    16607.295108638216
Rows/s:          16607.295108638216
raw latency :
  max:           13.860863ms 
  99.9th:        4.947967ms 
  99th:          2.424831ms 
  95th:          1.835007ms 
  90th:          1.703935ms 
  median:        786.431µs 
  mean:          959.485µs
c-o fixed latency :
  max:           13.860863ms 
  99.9th:        4.947967ms 
  99th:          2.424831ms 
  95th:          1.835007ms 
  90th:          1.703935ms 
  median:        786.431µs 
  mean:          961.792µs
  
master:
-----------------------------

 Results
Time (avg):      1m59.988254027s
Total ops:       1967868
Total rows:      1967868
Operations/s:    16400.505326073064
Rows/s:          16400.505326073064
raw latency :
  max:           16.842751ms 
  99.9th:        4.784127ms 
  99th:          2.457599ms 
  95th:          1.835007ms 
  90th:          1.671167ms 
  median:        819.199µs 
  mean:          971.348µs
c-o fixed latency :
  max:           16.842751ms 
  99.9th:        4.784127ms 
  99th:          2.490367ms 
  95th:          1.835007ms 
  90th:          1.671167ms 
  median:        851.967µs 
  mean:          973.814µs

```

Interesting observation though is that when introducing "disturbance" (i.e. a few extra processes at 100% cpu on some cores), master seems to have an edge, probably because it throws its IO in a pile at the kernel, and lets him deal with it, whereas we now (in this PR) tend to more wait for one set of IO to complete, and thus might be more susceptible to scheduling  delays causing latencies directly in the write path.